+++ src/Translator-test.js
@@ -82,7 +82,11 @@ var englishTrans = new Translator();
 englishTrans.setLanguage(['en_US']);
 englishTrans.use(NAMESPACE);
 
-testAsyncMulti("Translator - fill in parameter", [
+var germanTrans = new Translator();
+germanTrans.setLanguage(['de_DE']);
+germanTrans.use(NAMESPACE);
+
+testAsyncMulti("Translator - message-format - fill in parameter", [
   function (test, expect) {
     englishTrans.ready(expect(function () {
       var result = englishTrans.get('weather.message', { weather: "rainy" });
@@ -91,23 +95,41 @@ testAsyncMulti("Translator - fill in parameter", [
   }
 ]);
 
-testAsyncMulti("Translator - conditions", [
+testAsyncMulti("Translator - message-format - combine plural and select", [
   function (test, expect) {
     englishTrans.ready(expect(function () {
-      var result = englishTrans.get('fried_count', { friends: 0 });
-      test.equal(result, "You have no friends!");
+      var result = englishTrans.get('item_add', { count: 1, gender: 'female' });
+      test.equal(result, "She added one item!\n");
     }));
   },
   function (test, expect) {
     englishTrans.ready(expect(function () {
-      var result = englishTrans.get('fried_count', { friends: 1 });
-      test.equal(result, "You have a friend!");
+      var result = englishTrans.get('item_add', { count: 5, gender: 'female' });
+      test.equal(result, "She added 5 items!\n");
     }));
   },
   function (test, expect) {
     englishTrans.ready(expect(function () {
-      var result = englishTrans.get('fried_count', { friends: 5 });
-      test.equal(result, "You have 5 friends!");
+      var result = englishTrans.get('item_add', { count: 1, gender: 'male' });
+      test.equal(result, "He added one item!\n");
+    }));
+  },
+  function (test, expect) {
+    englishTrans.ready(expect(function () {
+      var result = englishTrans.get('item_add', { count: 5, gender: 'male' });
+      test.equal(result, "He added 5 items!\n");
+    }));
+  },
+  function (test, expect) {
+    englishTrans.ready(expect(function () {
+      var result = englishTrans.get('item_add', { count: 1, gender: null });
+      test.equal(result, "They added one item!\n");
+    }));
+  },
+  function (test, expect) {
+    englishTrans.ready(expect(function () {
+      var result = englishTrans.get('item_add', { count: 5, gender: null });
+      test.equal(result, "They added 5 items!\n");
     }));
   }
 ]);