+++ src/message-format/select-postprocess.js
@@ -1,0 +1,10 @@
+MessageFormatPostprocess.select = function (object, data) {
+  var parameter = data.parameters[object.name];
+  
+  var value = object.hash[parameter];
+  if (value == null) {
+    value = object.hash["other"]; // other also as fallback
+  }
+  
+  return Translator._messageFormatFilter(value, data); // recursion
+};